fix(CommentTeaser): use styleguide component and fix timestamp links
